move deprecated definitions to runfiles.go
