+++ go/tools/bazel/runfiles.go
@@ -248,6 +248,26 @@ func RunfilesPath() (string, error) {
 	return filepath.Join(runfiles.dir, runfiles.workspace), nil
 }
 
+// EnterRunfiles locates the directory under which a built binary can find its data dependencies
+// using relative paths, and enters that directory.
+//
+// "workspace" indicates the name of the current project, "pkg" indicates the relative path to the
+// build package that contains the binary target, "binary" indicates the basename of the binary
+// searched for, and "cookie" indicates an arbitrary data file that we expect to find within the
+// runfiles tree.
+//
+// DEPRECATED: use RunfilesPath instead.
+func EnterRunfiles(workspace string, pkg string, binary string, cookie string) error {
+	runfiles, ok := findRunfiles(workspace, pkg, binary, cookie)
+	if !ok {
+		return fmt.Errorf("cannot find runfiles tree")
+	}
+	if err := os.Chdir(runfiles); err != nil {
+		return fmt.Errorf("cannot enter runfiles tree: %v", err)
+	}
+	return nil
+}
+
 var runfiles = struct {
 	once, listOnce sync.Once
 
@@ -368,4 +388,38 @@ func initRunfiles() {
 		}
 		sort.Strings(runfiles.workspaces)
 	}
+}
+
+// getCandidates returns the list of all possible "prefix/suffix" paths where there might be an
+// optional component in-between the two pieces.
+//
+// This function exists to cope with issues #1239 because we cannot tell where the built Go
+// binaries are located upfront.
+//
+// DEPRECATED: only used by EnterRunfiles.
+func getCandidates(prefix string, suffix string) []string {
+	candidates := []string{filepath.Join(prefix, suffix)}
+	if entries, err := ioutil.ReadDir(prefix); err == nil {
+		for _, entry := range entries {
+			candidate := filepath.Join(prefix, entry.Name(), suffix)
+			candidates = append(candidates, candidate)
+		}
+	}
+	return candidates
+}
+
+// findRunfiles locates the directory under which a built binary can find its data dependencies
+// using relative paths.
+//
+// DEPRECATED: only used by EnterRunfiles.
+func findRunfiles(workspace string, pkg string, binary string, cookie string) (string, bool) {
+	candidates := getCandidates(filepath.Join("bazel-bin", pkg), filepath.Join(binary+".runfiles", workspace))
+	candidates = append(candidates, ".")
+
+	for _, candidate := range candidates {
+		if _, err := os.Stat(filepath.Join(candidate, cookie)); err == nil {
+			return candidate, true
+		}
+	}
+	return "", false
 }
